added tests and refactored
- replaced symlink creation code in LinuxFileUtil
- added PppAuthSecrets tests, further refactoring and added a bug note
- fix for #1451, test update

+++ kura/org.eclipse.kura.net.admin/src/main/java/org/eclipse/kura/net/admin/util/LinuxFileUtil.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2011, 2016 Eurotech and/or its affiliates
+ * Copyright (c) 2011, 2017 Eurotech and/or its affiliates
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -12,8 +12,9 @@
 package org.eclipse.kura.net.admin.util;
 
 import java.io.File;
+import java.nio.file.Files;
+import java.nio.file.Paths;
 
-import org.eclipse.kura.core.linux.util.LinuxProcessUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -37,24 +38,13 @@ public class LinuxFileUtil {
         s_logger.debug("Creating symbolic link from " + targetFile + " to " + sourceFile);
 
         File f = new File(targetFile);
-        if (f.exists()) {
-            if (deleteOldLink) {
-                s_logger.debug("Deleting existing link");
-                f.delete();
-            }
+        if (f.exists() && deleteOldLink) {
+            s_logger.debug("Deleting existing link");
+            f.delete();
         }
 
         if (!f.exists()) {
-            StringBuffer buf = new StringBuffer("ln -s");
-            buf.append(' ');
-            buf.append(sourceFile);
-            buf.append(' ');
-            buf.append(targetFile);
-
-            int ret = LinuxProcessUtil.start(buf.toString());
-            if (ret != 0) {
-                throw new Exception("error executing command - " + buf.toString());
-            }
+            Files.createSymbolicLink(Paths.get(targetFile), Paths.get(sourceFile));
         }
     }
 }
